m3_proj_tf-bq-dk_nytaxi_yellow/scripts/pyproject.toml
[project]
name = "scripts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-cloud-bigquery>=3.40.0",
    "google-cloud-storage>=3.9.0",
    "pandas>=3.0.0",
    "pyarrow>=23.0.0",
    "requests>=2.32.5",
]
